snapshots: move encoding modules to common
